# docs/swagger.yaml
basePath: /
definitions:
  budget_plan.BudgetPlanDTO:
    properties:
      id:
        type: integer
      iscurrent:
        type: boolean
      items:
        items:
          $ref: '#/definitions/budget_plan.ItemDTO'
        type: array
      name:
        type: string
    type: object
  budget_plan.ItemDTO:
    properties:
      color:
        type: string
      icon:
        type: string
      id:
        type: integer
      name:
        type: string
      weeklyDuration:
        type: integer
      weeklyOccurrences:
        type: integer
    type: object
...
host: localhost:8181
info:
  contact: {}
  description: Time planning and tracking API
  title: Klokku API
  version: "0.9"
paths:
  /api/budgetplan:
    get:
      description: Get a list of all budget plans for the current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/budget_plan.BudgetPlanDTO'
            type: array
        "403":
          description: User not found
          schema:
            type: string
      security:
      - XUserId: []
      summary: List all budget plans
      tags:
      - BudgetPlan
    post:
      consumes:
      - application/json
      description: Create a new budget plan with the provided details
      parameters:
      - description: Budget Plan
        in: body
        name: plan
        required: true
        schema:
          $ref: '#/definitions/budget_plan.BudgetPlanDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/budget_plan.BudgetPlanDTO'
        "400":
          description: Bad Request
          schema:
            type: string
        "403":
          description: User not found
          schema:
            type: string
      security:
      - XUserId: []
      summary: Create a new budget plan
      tags:
      - BudgetPlan
  /api/budgetplan/{planId}:
    delete:
      description: Delete a budget plan by ID
      parameters:
      - description: Budget Plan ID
        in: path
        name: planId
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            type: string
        "403":
          description: User not found
          schema:
            type: string
        "404":
          description: Plan Not Found
          schema:
            type: string
      security:
      - XUserId: []
      summary: Delete a budget plan
      tags:
      - BudgetPlan
    get:
      description: Retrieve a specific budget plan with all its items
      parameters:
      - description: Budget Plan ID
        in: path
        name: planId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/budget_plan.BudgetPlanDTO'
        "400":
          description: Bad Request
          schema:
            type: string
        "403":
          description: User not found
          schema:
            type: string
        "404":
          description: Plan Not Found
          schema:
            type: string
      security:
      - XUserId: []
      summary: Get a budget plan by ID
      tags:
      - BudgetPlan
    put:
      consumes:
      - application/json
      description: Update a budget plan by ID
      parameters:
      - description: Budget Plan ID
        in: path
        name: planId
        required: true
        type: integer
      - description: Budget Plan
        in: body
        name: plan
        required: true
        schema:
          $ref: '#/definitions/budget_plan.BudgetPlanDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/budget_plan.BudgetPlanDTO'
        "400":
          description: Bad Request
          schema:
            type: string
        "403":
          description: User not found
          schema:
            type: string
        "404":
          description: Plan Not Found
          schema:
            type: string
      security:
      - XUserId: []
      summary: Update an existing budget plan
      tags:
      - BudgetPlan
  /api/budgetplan/{planId}/item:
    post:
      consumes:
      - application/json
      description: Register a new budget item within a specific budget plan
      parameters:
      - description: Budget Plan ID
        in: path
        name: planId
        required: true
        type: integer
      - description: Budget Item
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/budget_plan.ItemDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/budget_plan.ItemDTO'
        "400":
          description: Bad Request
          schema:
            type: string
        "403":
          description: User not found
          schema:
            type: string
      security:
      - XUserId: []
      summary: Add a new budget item to a plan
      tags:
      - BudgetItem
  /api/budgetplan/{planId}/item/{itemId}:
    delete:
      description: Remove a budget item from a plan
      parameters:
      - description: Budget Plan ID
        in: path
        name: planId
        required: true
        type: integer
      - description: Budget Item ID
        in: path
        name: itemId
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            type: string
        "403":
          description: User not found
          schema:
            type: string
        "404":
          description: Item Not Found
          schema:
            type: string
      security:
      - XUserId: []
      summary: Delete a budget item
      tags:
      - BudgetItem
    put:
      consumes:
      - application/json
      description: Update an existing budget item within a plan
      parameters:
      - description: Budget Plan ID
        in: path
        name: planId
        required: true
        type: integer
      - description: Budget Item ID
        in: path
        name: itemId
        required: true
        type: integer
      - description: Budget Item
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/budget_plan.ItemDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/budget_plan.ItemDTO'
        "400":
          description: Bad Request
          schema:
            type: string
        "403":
          description: User not found
          schema:
            type: string
      security:
      - XUserId: []
      summary: Update a budget item
      tags:
      - BudgetItem
  /api/budgetplan/{planId}/item/{itemId}/position:
    put:
      consumes:
      - application/json
      description: Move a budget item to a specific position in the list
      parameters:
      - description: Budget Plan ID
        in: path
        name: planId
        required: true
        type: integer
      - description: Budget Item ID
        in: path
        name: itemId
        required: true
        type: integer
      - description: Position details
        in: body
        name: position
        required: true
        schema:
          properties:
            id:
              type: integer
            precedingId:
              type: integer
          type: object
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            type: string
        "403":
          description: User not found
          schema:
            type: string
        "404":
          description: Item Not Found
          schema:
            type: string
      security:
      - XUserId: []
      summary: Set position of a budget item
      tags:
      - BudgetItem
...
securityDefinitions:
  XUserId:
    description: User ID header required for authentication
    in: header
    name: X-User-Id
    type: apiKey
swagger: "2.0"